Separating editor ui files into their own files.
