+++ editor/CMakeLists.txt
@@ -36,6 +36,11 @@ add_library(${PROJECT_NAME} STATIC
         src/core/ui/imgui/imgui_helper.cpp
         src/core/ui/imgui/imgui_styler.cpp
         src/core/ui/imgui/imgui_window_renderer.cpp
+        src/core/ui/views/opened_project/menu_bar_ui.cpp
+        src/core/ui/views/opened_project/windows/asset_browser_ui.cpp
+        src/core/ui/views/opened_project/windows/details_ui.cpp
+        src/core/ui/views/opened_project/windows/scene_outliner_ui.cpp
+        src/core/ui/views/opened_project/windows/scene_view_ui.cpp
         src/core/utils/console_logger.cpp
         src/core/utils/process_runner/process_runner.cpp
         src/core/utils/process_runner/process_context.cpp
