+++ editor/src/core/ui/views/opened_project/menu_bar_ui.h
@@ -1,0 +1,9 @@
+#pragma once
+
+#include "../../imgui/imgui_helper.h"
+
+namespace OpenedProjectUI {
+namespace MenuBar {
+ImGuiHelper::MenuBar GetMenuBar();
+} // namespace MenuBar
+} // namespace OpenedProjectUI
