+++ editor/src/core/ui/views/opened_project/windows/console_ui.h
@@ -1,0 +1,1 @@
+#pragma once
