+++ editor/src/core/ui/views/opened_project/windows/details_ui.h
@@ -1,0 +1,9 @@
+#pragma once
+
+#include "../../../imgui/imgui_helper.h"
+
+namespace OpenedProjectUI {
+namespace Windows {
+ImGuiHelper::Window GetDetailsWindow();
+} // namespace Windows
+} // namespace OpenedProjectUI
