+++ editor/src/core/ui/views/opened_project/windows/scene_view_ui.h
@@ -1,0 +1,9 @@
+#pragma once
+
+#include "../../../imgui/imgui_helper.h"
+
+namespace OpenedProjectUI {
+namespace Windows {
+ImGuiHelper::Window GetSceneViewWindow();
+} // namespace Windows
+} // namespace OpenedProjectUI
